requirements.txt
flask==2.3.3
flask-cors==4.0.0
gitpython==3.1.41
python-dotenv==1.0.0
schedule==1.2.1
Werkzeug==2.3.7 